requirements.txt
APScheduler==3.6.3
asn1crypto==0.24.0
Authlib==0.12.1
awsebcli==3.10.3
boto==2.48.0
botocore==1.5.82
braintree==3.25.0
cachetools==3.1.1
cement==2.8.2
cffi==1.12.3
colorama==0.3.7
cryptography==2.7
defusedxml==0.5.0
dj-database-url==0.4.0
Django==1.11.4
django-background-tasks==1.2.5
django-compat==1.0.15
django-storages==1.6.3
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
gspread==3.3.0
gunicorn==19.4.5
httplib2==0.14.0
jmespath==0.9.3
oauth2client==4.1.3
oauthlib==3.1.0
pathspec==0.5.0
Pillow==5.4.1
psycopg2==2.7.4
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
PyJWT==1.4.0
pytesseract==0.3.0
python-dateutil==2.6.1
python-http-client==3.2.6
python-social-auth==0.2.14
python3-openid==3.0.10
pytz==2017.2
PyYAML==3.12
requests==2.9.1
requests-oauthlib==1.3.0
rsa==4.0
semantic-version==2.5.0
sendgrid==6.2.0
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
tabulate==0.7.5
termcolor==1.1.0
tzlocal==2.0.0
uritemplate==3.0.0
websocket-client==0.44.0
whitenoise==3.0
